creating a scaffold for FIP objects
Tidied up a lot of WIP global-scoped code, and pulled it into the FIP object instance.
The idea is to bundle the information in the system hierarchy file into a series of FIP objects, and not handle interacting lists of objects of different types...should be easier and clearer than the FIP_experimental attempt.